requirements.txt
dash==2.9.3
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
plotly==5.9.0
python_dateutil==2.8.2
gunicorn==20.0.0
pyarrow==12.0.0
Werkzeug==2.2.2
xlsxwriter==3.1.2
Flask==2.2.2
Flask-Caching==2.0.2
click==8.0.4
Jinja2==3.1.2
itsdangerous==2.1.2
python-dotenv==1.0.0
